pyproject.toml
[project]
name = "rentcodedjangoapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django-ninja>=1.3.0",
    "django>=5.1.4",
    "django-ninja-extra>=0.21.8",
    "pydantic>=2.10.3",
    "pillow>=11.0.0",
    "ruff>=0.8.3",
    "django-debug-toolbar>=4.4.6",
    "ipython>=8.30.0",
]
